feat(handdecomp): port UnitList_DrawPageHeader, LoadUnit_MoveToPosition → matching C
screen NEAR s8-param sign-ext (lsr->asr); int param + (s8) cast. matching-C 7645→7647.

+++ include/event.h
@@ -417,7 +417,7 @@ u8 Event2A_MoveToChapter(struct EventEngineProc * proc);
 // ??? EventLoadUnitWithMovement(???);
 void EventLoadUnitSliently(const struct UnitDefinition* def, s16 count, u8 param);
 void LoadUnit_0(const struct UnitDefinition *, u16, s8, s8);
-void LoadUnit_MoveToPosition(struct Unit*, const struct UnitDefinition*, u16, s8);
+void LoadUnit_MoveToPosition(struct Unit*, const struct UnitDefinition*, u16, int);
 // ??? GetUnitDefinitionFormEventScr(???);
 u8 Event2B_ConfigLoadUnit(struct EventEngineProc * proc);
 u8 Event2C_LoadUnits(struct EventEngineProc * proc);
